Fix DB init error: Running by docker/sql/run-init.sh / Auto create Database , table -> insert data to room and timeslot/ Adding folder login,room in dto and service to store source code
