+++ docker/sql/init_core.sql
@@ -1,43 +1,47 @@
-/* 0) Database */
+:on error exit
+
+-- 0) Create DB if needed
 IF DB_ID('bookingDB') IS NULL
 BEGIN
   PRINT 'Creating database [bookingDB]...';
   CREATE DATABASE [bookingDB];
 END
-ELSE
-  PRINT 'Database [bookingDB] already exists.';
+GO
 
-/* 1) Login & User */
+-- 0.1) Wait until DB is ONLINE
+DECLARE @i int = 0;
+WHILE DB_ID('bookingDB') IS NULL 
+   OR EXISTS (SELECT 1 FROM sys.databases WHERE name = 'bookingDB' AND state_desc <> 'ONLINE')
+BEGIN
+  SET @i += 1;
+  PRINT CONCAT('Waiting bookingDB ONLINE... (', @i, ')');
+  WAITFOR DELAY '00:00:01';
+END
+GO
+
+-- 1) Login
 USE [master];
 IF NOT EXISTS (SELECT 1 FROM sys.sql_logins WHERE name = 'booking_app')
 BEGIN
   PRINT 'Creating login [booking_app]...';
   EXEC('CREATE LOGIN [booking_app] WITH PASSWORD = N''__APPPASS__'', CHECK_POLICY = OFF, CHECK_EXPIRATION = OFF;');
 END
-ELSE
-  PRINT 'Login [booking_app] already exists.';
+GO
 
+-- 2) User & role
 USE [bookingDB];
 IF NOT EXISTS (SELECT 1 FROM sys.database_principals WHERE name = 'booking_app')
-BEGIN
-  PRINT 'Creating user [booking_app]...';
   CREATE USER [booking_app] FOR LOGIN [booking_app];
-END
-ELSE
-  PRINT 'User [booking_app] already exists.';
-
 IF NOT EXISTS (
   SELECT 1
   FROM sys.database_role_members rm
   JOIN sys.database_principals r ON rm.role_principal_id = r.principal_id AND r.name='db_owner'
   JOIN sys.database_principals u ON rm.member_principal_id = u.principal_id AND u.name='booking_app'
 )
-BEGIN
   ALTER ROLE db_owner ADD MEMBER [booking_app];
-END
+GO
 
 /* 2) Tables (idempotent) */
--- rooms
 IF OBJECT_ID('dbo.rooms','U') IS NULL BEGIN
   PRINT 'Creating table dbo.rooms';
   CREATE TABLE dbo.rooms(
@@ -54,7 +58,6 @@ IF OBJECT_ID('dbo.rooms','U') IS NULL BEGIN
   CREATE UNIQUE INDEX uk_rooms_code ON dbo.rooms(code);
 END
 
--- time_slots
 IF OBJECT_ID('dbo.time_slots','U') IS NULL BEGIN
   PRINT 'Creating table dbo.time_slots';
   CREATE TABLE dbo.time_slots(
@@ -66,7 +69,6 @@ IF OBJECT_ID('dbo.time_slots','U') IS NULL BEGIN
   CREATE UNIQUE INDEX uk_time_slots_code ON dbo.time_slots(slot_code);
 END
 
--- reservations
 IF OBJECT_ID('dbo.reservations','U') IS NULL BEGIN
   PRINT 'Creating table dbo.reservations';
   CREATE TABLE dbo.reservations(
@@ -81,4 +83,3 @@ IF OBJECT_ID('dbo.reservations','U') IS NULL BEGIN
     CONSTRAINT uk_res_unique UNIQUE(room_id, slot_id, [date])
   );
 END
-PRINT 'Database initialization script completed.';
