+++ src/main/java/com/example/lc2_booking_room/dto/login/RequestOtpDto.java
@@ -1,21 +1,14 @@
-package com.example.lc2_booking_room.dto;
+package com.example.lc2_booking_room.dto.login;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public class VerifyOtpDto {
+public class RequestOtpDto {
     @NotBlank @Email
     private String email;
 
-    @NotBlank
-    private String otp;
-
-    public VerifyOtpDto() {}
+    public RequestOtpDto() {}
 
     public String getEmail() { return email; }
     public void setEmail(String email) { this.email = email; }
-
-    public String getOtp() { return otp; }
-    public void setOtp(String otp) { this.otp = otp; }
 }
-
