+++ src/main/java/com/example/lc2_booking_room/dto/login/TuCheckDto.java
@@ -1,16 +1,16 @@
-package com.example.lc2_booking_room.dto;
+package com.example.lc2_booking_room.dto.login;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public class TuCheckRequest {
+public class TuCheckDto {
     @NotBlank
     private String userName;
 
     @NotBlank @Email
     private String email;
 
-    public TuCheckRequest() {}
+    public TuCheckDto() {}
 
     public String getUserName() { return userName; }
     public void setUserName(String userName) { this.userName = userName; }
