+++ src/main/java/com/example/lc2_booking_room/dto/login/VerifyOtpDto.java
@@ -1,14 +1,21 @@
-package com.example.lc2_booking_room.dto;
+package com.example.lc2_booking_room.dto.login;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public class RequestOtpDto {
+public class VerifyOtpDto {
     @NotBlank @Email
     private String email;
 
-    public RequestOtpDto() {}
+    @NotBlank
+    private String otp;
+
+    public VerifyOtpDto() {}
 
     public String getEmail() { return email; }
     public void setEmail(String email) { this.email = email; }
+
+    public String getOtp() { return otp; }
+    public void setOtp(String otp) { this.otp = otp; }
 }
+
