+++ src/main/java/com/example/lc2_booking_room/security/JwtAuthenticationFilter.java
@@ -1,6 +1,5 @@
 package com.example.lc2_booking_room.security;
 
-import com.example.lc2_booking_room.service.JwtService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
@@ -10,6 +9,8 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
+
+import com.example.lc2_booking_room.service.login.JwtService;
 
 import java.io.IOException;
 import java.util.List;
